+++ word_projections/errors.py
@@ -1,0 +1,5 @@
+class WordNotFound400(Exception):
+    pass
+
+class AxisNotFound400(Exception):
+    pass
